Module Conduit_lwt_tls

Implementation of the TLS support (according ocaml-tls) with conduit-lwt.

This implementation is a specialization of conduit-tls with conduit-lwt. Underlying protocol or service can be anything into the scope of conduit-lwt.

For more details about behaviours, you should look into conduit-tls.

type 'flow protocol_with_tls

val underlying : 'flow protocol_with_tls -> 'flow

underlying tls_flow returns the underlying flow used with TLS.

val handshake : 'flow protocol_with_tls -> bool

handshake flow returns true if the handshake is processing. Otherwise, it returns false.

val protocol_with_tls : ('edn, 'flow) Conduit_lwt.protocol -> ('edn * Tls.Config.client, 'flow protocol_with_tls) Conduit_lwt.protocol

type 'service service_with_tls

val service_with_tls : ('cfg, 't, 'flow) Conduit_lwt.Service.service -> ('edn, 'flow protocol_with_tls) Conduit_lwt.protocol -> ('cfg * Tls.Config.server, 't service_with_tls, 'flow protocol_with_tls) Conduit_lwt.Service.service

module TCP : sig ... end